backend/requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
# RabbitMQ y Celery
pika==1.3.2
celery==5.3.4
django-celery-results==2.5.1
# Swagger
drf-spectacular==0.26.3
drf-spectacular-sidecar==2024.4.1
# PostgreSQL
psycopg2-binary==2.9.9
# CORS para frontend
django-cors-headers==4.3.1
python-decouple==3.8